Commit changes for the RIL in the Motorla Triumph.  Note that these are enabled only if the Triumph variable is declared in build.prop - you need to declare the following in system.prop under the build folder:
ro.telephony.ril_nid=97
ro.telephony.ril_prl=1115
ro.telephony.ril_class=Triumph
ro.ril.def.preferred.network=4
ro.telephony.default_network=4
rild.libpath=/system/lib/libril-qc-1.so
rild.libargs=-d /dev/smd0

The PRL is a fake but must be declared and the ril_class must be declared
or the code will not properly translate the state changes for the Triumph;
if it is missing the special case code will not execute.

Qualcomm/Motorola, in short, are sending back a "SIM_NOT_READY" status and
NEVER send back the proper "NV_READY" state transition.  This prevents the
base code from ever polling for the subscription data, and without that
the RIL code does not transition to the online state.  SIM_NOT_READY
also causes a teardown of the radio and the inappropriate switch to
GSM mode (since it's more-or-less a declaration that you have a SIM,
but it's not yet available.  In addition the radio code sends back four
fields (like a Samsung) rather than five.  There are a couple of other
minor tweeks as well.

libril-qc-1.so and libril.so must come from the 2.2 Froyo Triumph build.